airflow/etl/requirements.txt
requests==2.32.3
psycopg2-binary==2.9.10
python-decouple==3.8
python-dateutil==2.9.0.post0